join화면 순서 교체
